Removed old command
* Added "auto-join" channel feature.
